# KDESKTOP-930 Check for getRights method performance at startup and if…