--- a/src/libcommonserver/io/iohelper.h
+++ b/src/libcommonserver/io/iohelper.h
@@ -54,16 +54,9 @@ struct IoHelper {
                 SyncPath _directoryPath;
                 std::filesystem::recursive_directory_iterator _dirIterator;
         };
+        IoHelper() = default;
 
-    public:
-        IoHelper(){};
-
-        inline static void setLogger(log4cplus::Logger logger) { _logger = logger; }
-
-#ifdef _WIN32
-        static int _getAndSetRightsMethod;
-#endif
-
+        inline static void setLogger(const log4cplus::Logger &logger) { _logger = logger; }
         static IoError stdError2ioError(int error) noexcept;
         static IoError stdError2ioError(const std::error_code &ec) noexcept;
         static IoError posixError2ioError(int error) noexcept;
@@ -337,6 +330,7 @@ struct IoHelper {
         static bool createJunction(const std::string &data, const SyncPath &path, IoError &ioError) noexcept;
         static bool readJunction(const SyncPath &path, std::string &data, SyncPath &targetPath, IoError &ioError) noexcept;
         static bool createJunctionFromPath(const SyncPath &targetPath, const SyncPath &path, IoError &ioError) noexcept;
+        static TRUSTEE &getTrustee();
 #endif
         //! Checks if the item indicated by the specified path is dehydrated.
         /*!
@@ -408,6 +402,10 @@ struct IoHelper {
 
 #ifdef _WIN32
         static bool _setRightsWindowsApiInheritance;  // For windows tests only
+        static int _getAndSetRightsMethod;
+        static std::unique_ptr<BYTE[]> _psid;
+        static TRUSTEE _trustee;
+        static void initRightsWindowsApi();
 #endif
 };