--- a/src/libcommonserver/utility/utility_win.cpp
+++ b/src/libcommonserver/utility/utility_win.cpp
@@ -145,7 +145,6 @@ static bool init_private() {
         }
     }
     auto end = std::chrono::high_resolution_clock::now();
-
     auto duration = std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count();
 
     /* Averrage times for 10 calls to getRights with windows API: